pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skill-forge"
version = "0.1.0"
description = "A CLI demo showcasing Agent Skills with progressive disclosure"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer>=0.12.0",
    "python-frontmatter>=1.0.0",
    "anthropic>=0.34.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
skill-forge = "skill_forge.cli:main"

[tool.setuptools]
packages = ["skill_forge"]

